more fixes

+++ lib_fun/ft_numlen.c
@@ -6,13 +6,13 @@
 /*   By: elhampto <elhampto@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/23 16:31:37 by elhampto          #+#    #+#             */
-/*   Updated: 2019/07/24 23:30:23 by elhampto         ###   ########.fr       */
+/*   Updated: 2019/07/29 22:52:07 by elhampto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inclu/ft_printf.h"
 
-int				ft_numlen(int num)
+int				ft_numlen(uint64_t num)
 {
 	int			len;
